requirements.txt
Django==1.10
djangorestframework==3.4.1
factory-boy==2.7.0
fake-factory==0.6.0
python-dateutil==2.5.3
six==1.10.0
